added ssl user flag

+++ server-init/init.sh
@@ -2,6 +2,17 @@
 
 echo "Hi!\nInitiating UoS3 server config\n"
 
+SSL_FLAG=0
+if [ ! -z $1 ]; then
+  [ "$1" = "-ssl" ] && {
+    SSL_FLAG=1
+    echo "SSL files generation option has been specified."
+  } || {
+    echo "Flag $1 not recognised. Aborting..."
+    exit 1
+  }
+fi
+
 USERS=$( cat "`dirname $0`/SECRETS/AUTH_USERS" )
 [ -z "$USERS" ] && {
   echo "FATAL: You forgot to add the list of authorised users! Exiting..."
@@ -12,12 +23,14 @@ SSL_DIR=$( cat "`dirname $0`/SECRETS/SSL_DIR" )
   echo "FATAL: You forgot to add the secret UoS3 SSL dir path! Exiting..."
   exit 1
 }
-[ ! -f "`dirname $0`/SECRETS/SSL_STRONG_CONF" ] && {
-  SSL_STRONG_CONF=""
-  echo "You have not specified strong SSL configs. Leaving at weak SSL for now."
-} || {
-  SSL_STRONG_CONF="`dirname $0`/SECRETS/SSL_STRONG_CONF"
-}
+if [ $SSL_FLAG = 1 ]; then
+  [ ! -f "`dirname $0`/SECRETS/SSL_STRONG_CONF" ] && {
+    SSL_STRONG_CONF=""
+    echo "You have not specified strong SSL configs. Leaving at weak SSL for now."
+  } || {
+    SSL_STRONG_CONF="`dirname $0`/SECRETS/SSL_STRONG_CONF"
+  }
+fi
 
 if ! which nginx > /dev/null 2>&1; then
   echo "nginx is not installed! Installing...\n"
@@ -27,12 +40,14 @@ if ! which nginx > /dev/null 2>&1; then
     exit 1
   fi
 fi
-if ! which openssl > /dev/null 2>&1; then
-  echo "OpenSSL required for secure connections. Installing...\n"
-  sudo apt-get update && sudo apt-get -y install openssl
+if [ $SSL_FLAG = 1 ]; then
   if ! which openssl > /dev/null 2>&1; then
-    echo "\nFATAL: OpenSSL is not installed! Exiting..."
-    exit 1
+    echo "OpenSSL required for secure connections. Installing...\n"
+    sudo apt-get update && sudo apt-get -y install openssl
+    if ! which openssl > /dev/null 2>&1; then
+      echo "\nFATAL: OpenSSL is not installed! Exiting..."
+      exit 1
+    fi
   fi
 fi
 if ! which htpasswd > /dev/null 2>&1; then
@@ -76,38 +91,42 @@ for user in $USERS; do
   fi
 done
 
-if [ ! -f "${SSL_DIR}/private/uos3.key" -o ! -f "${SSL_DIR}/certs/uos3.crt" ]; then
-  echo "SSL cert is required. Creating..."
-  echo "This step is important. Confirm that you know what you are doing. (y/N)"
-  read PROMPT
-  [ "$PROMPT" = "y" -o "$PROMPT" = "Y" ] && {
-    sudo openssl req -x509 -nodes -days 365 -newkey rsa:2048 \
-      -keyout ${SSL_DIR}/private/uos3.key -out ${SSL_DIR}/certs/uos3.crt
-    sudo openssl dhparam -out ${SSL_DIR}/dhparam.pem 4096
-  } || {
-    echo "Aborted. Goodbye."
-    exit 1
-  }
-  [ ! -f "${SSL_DIR}/private/uos3.key" -o ! -f "${SSL_DIR}/certs/uos3.crt" ] && {
-    echo "\nFATAL: Cannot create SSL cert! Exiting..."; exit 1;
-  } || echo "\nCreated\n"
-else
-  echo "Found existing SSL cert\n"
+if [ $SSL_FLAG = 1 ]; then
+  if [ ! -f "${SSL_DIR}/private/uos3.key" -o ! -f "${SSL_DIR}/certs/uos3.crt" ]; then
+    echo "SSL cert is required. Creating..."
+    echo "This step is important. Confirm that you know what you are doing. (y/N)"
+    read PROMPT
+    [ "$PROMPT" = "y" -o "$PROMPT" = "Y" ] && {
+      sudo openssl req -x509 -nodes -days 365 -newkey rsa:2048 \
+        -keyout ${SSL_DIR}/private/uos3.key -out ${SSL_DIR}/certs/uos3.crt
+      sudo openssl dhparam -out ${SSL_DIR}/dhparam.pem 4096
+    } || {
+      echo "Aborted. Goodbye."
+      exit 1
+    }
+    [ ! -f "${SSL_DIR}/private/uos3.key" -o ! -f "${SSL_DIR}/certs/uos3.crt" ] && {
+      echo "\nFATAL: Cannot create SSL cert! Exiting..."; exit 1;
+    } || echo "\nCreated\n"
+  else
+    echo "Found existing SSL cert\n"
+  fi
 fi
 
-echo "Remaking SSL snippets..."
-[ -d ${SSL_DIR}/snippets ] && sudo rm -rfv ${SSL_DIR}/snippets
-sudo mkdir -pv ${SSL_DIR}/snippets
-SNIP_CRT_POINT="${SSL_DIR}/snippets/uos3-cert-pointer.conf"
-sudo touch $SNIP_CRT_POINT
-echo "ssl_certificate ${SSL_DIR}/certs/uos3.crt" | sudo tee -a $SNIP_CRT_POINT
-echo "ssl_certificate_key ${SSL_DIR}/private/uos3.key" | sudo tee -a $SNIP_CRT_POINT
-[ ! -z "$SSL_STRONG_CONF" ] && {
-  SNIP_STRONG_SSL="${SSL_DIR}/snippets/uos3-strong-ssl.conf"
-  sudo touch $SNIP_STRONG_SSL
-  cat $SSL_STRONG_CONF | sudo tee -a $SNIP_STRONG_SSL
-}
-echo "SSL snippets remade\n"
+if [ $SSL_FLAG = 1 ]; then
+  echo "Remaking SSL snippets..."
+  [ -d ${SSL_DIR}/snippets ] && sudo rm -rfv ${SSL_DIR}/snippets
+  sudo mkdir -pv ${SSL_DIR}/snippets
+  SNIP_CRT_POINT="${SSL_DIR}/snippets/uos3-cert-pointer.conf"
+  sudo touch $SNIP_CRT_POINT
+  echo "ssl_certificate ${SSL_DIR}/certs/uos3.crt" | sudo tee -a $SNIP_CRT_POINT
+  echo "ssl_certificate_key ${SSL_DIR}/private/uos3.key" | sudo tee -a $SNIP_CRT_POINT
+  [ ! -z "$SSL_STRONG_CONF" ] && {
+    SNIP_STRONG_SSL="${SSL_DIR}/snippets/uos3-strong-ssl.conf"
+    sudo touch $SNIP_STRONG_SSL
+    cat $SSL_STRONG_CONF | sudo tee -a $SNIP_STRONG_SSL
+  }
+  echo "SSL snippets remade\n"
+fi
 
 echo "Adding user credentials to nginx config..."
